+++ js/src/components/LessonGrid.js
@@ -1,10 +1,10 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import Box from '@material-ui/core/Box';
 import makeStyles from '@material-ui/core/styles/makeStyles';
-import PropTypes from 'prop-types';
 
 const useStyles = makeStyles((theme) => ({
-  inner: {
+  innerDefault: {
     paddingLeft: theme.spacing(2),
     paddingRight: theme.spacing(2),
     [theme.breakpoints.up('sm')]: {
@@ -19,15 +19,30 @@ const useStyles = makeStyles((theme) => ({
       marginRight: 'auto',
     },
   },
+  innerIgnorePaddings: {
+    marginRight: theme.spacing(3.75),
+    [theme.breakpoints.down('md')]: {
+      marginRight: theme.spacing(2),
+    },
+  },
 }));
 
-const LessonGrid = ({ children }) => {
+const LessonGrid = ({ children, ignorePaddings }) => {
   const classes = useStyles();
-  return <Box className={classes.inner}>{children}</Box>;
+  return (
+    <Box className={!ignorePaddings ? classes.innerDefault : classes.innerIgnorePaddings}>
+      {children}
+    </Box>
+  );
 };
 
 LessonGrid.propTypes = {
   children: PropTypes.node,
+  ignorePaddings: PropTypes.bool,
+};
+
+LessonGrid.defaultProps = {
+  ignorePaddings: false,
 };
 
 export default LessonGrid;