+++ js/src/pages/lesson/ContentQuiz.js
@@ -1,20 +1,32 @@
 import React, { useState } from 'react';
 import Box from '@material-ui/core/Box';
 import Typography from '@material-ui/core/Typography';
+import Hidden from '@material-ui/core/Hidden';
 import LessonGrid from '@anu/components/LessonGrid';
 import ParagraphsWithQuiz from '@anu/components/ParagraphsWithQuiz';
 import { quizPropTypes } from '@anu/utilities/transform.quiz';
 import ContentNavigation from '@anu/components/ContentNavigation';
-import Hidden from '@material-ui/core/Hidden';
 import { lessonPropTypes } from '@anu/utilities/transform.lesson';
+import ContentTopNavigation from '@anu/components/TopContentNavigation';
+import { coursePropTypes } from '@anu/utilities/transform.course';
 
-const ContentQuiz = ({ quiz, nextLesson, prevLesson }) => {
+const ContentQuiz = ({ quiz, nextLesson, prevLesson, course }) => {
   const [isSubmitted, submitQuiz] = useState(!!quiz.isSubmitted);
   return (
-    <Box mt={[2, 2, 0]}>
+    <Box>
+      <ContentTopNavigation
+        isIntro={false}
+        sections={[]}
+        currentLesson={quiz}
+        nextLesson={nextLesson}
+        prevLesson={prevLesson}
+        currentIndex={0}
+        isEnabled={isSubmitted}
+        course={course}
+      />
       <LessonGrid>
         <Box display="flex" justifyContent="space-between" alignItems="center">
-          <Hidden smDown>
+          <Hidden mdUp>
             <Typography variant="h4" component="h1">
               {quiz.title}
             </Typography>
@@ -43,6 +55,8 @@ const ContentQuiz = ({ quiz, nextLesson, prevLesson }) => {
           prevLesson={prevLesson}
           currentIndex={0}
           isEnabled={isSubmitted}
+          ignorePaddings={true}
+          hideButtonsLabelsOnMobile={false}
         />
       )}
     </Box>
@@ -53,6 +67,7 @@ ContentQuiz.propTypes = {
   quiz: quizPropTypes,
   nextLesson: lessonPropTypes,
   prevLesson: lessonPropTypes,
+  course: coursePropTypes,
 };
 
 export default ContentQuiz;