+++ js/src/pages/lesson/NavigationMobile.js
@@ -1,11 +1,8 @@
 import React, { useState } from 'react';
-import Sticky from 'react-sticky-el';
 
 import Box from '@material-ui/core/Box';
-import Grid from '@material-ui/core/Grid';
 import Drawer from '@material-ui/core/Drawer';
 import Container from '@material-ui/core/Container';
-import Typography from '@material-ui/core/Typography';
 import CloseIcon from '@material-ui/icons/Close';
 import makeStyles from '@material-ui/core/styles/makeStyles';
 import DoubleArrowIcon from '@material-ui/icons/DoubleArrow';
@@ -17,22 +14,22 @@ import { getPwaSettings } from '@anu/utilities/settings';
 import DownloadCoursePopup from '@anu/components/DownloadCoursePopup';
 
 const useStyles = makeStyles((theme) => ({
-  sticky: {
-    zIndex: 4,
-  },
-  stickyContent: {
-    background: theme.palette.grey[200],
-    fontWeight: theme.typography.fontWeightBold,
-    color: theme.palette.common.black,
-    cursor: 'pointer',
-    borderBottom: '2px solid ' + theme.palette.common.black,
-  },
   chevron: ({ isVisible }) => ({
     transform: isVisible ? 'rotate(270deg)' : 'rotate(90deg)',
     transition: '.2s transform',
-    fontSize: '1rem',
-    float: 'right',
+    fontSize: '1.5rem',
   }),
+  openSection: {
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+    cursor: 'pointer',
+    height: theme.spacing(8),
+    marginLeft: -theme.spacing(3),
+    [theme.breakpoints.down('md')]: {
+      marginLeft: -theme.spacing(6),
+    },
+  },
   closeIcon: {
     cursor: 'pointer',
   },
@@ -44,21 +41,11 @@ const LessonNavigationMobile = ({ lesson, course }) => {
 
   return (
     <>
-      {/* Sticky lesson page title */}
-      <Sticky className={classes.sticky}>
-        <Container className={classes.stickyContent}>
-          <Box pt={2} pb={2} onClick={() => toggleVisibility(true)}>
-            <Grid container alignItems="center">
-              <Grid item xs={11}>
-                <Typography variant="subtitle2">{lesson.title}</Typography>
-              </Grid>
-              <Grid item xs={1}>
-                <DoubleArrowIcon className={classes.chevron} />
-              </Grid>
-            </Grid>
-          </Box>
-        </Container>
-      </Sticky>
+      <Container>
+        <Box onClick={() => toggleVisibility(true)} className={classes.openSection}>
+          <DoubleArrowIcon className={classes.chevron} />
+        </Box>
+      </Container>
 
       <Drawer anchor="bottom" open={isVisible} onClose={() => toggleVisibility(false)}>
         {/* Close button to hide the drawer */}