+++ js/src/utilities/transform.lesson.js
@@ -35,7 +35,6 @@ const transformLesson = (node, course) => {
           .getArrayValue(content, 'field_lesson_section_content')
           .map((paragraph) => transformParagraph(paragraph))
       ),
-    finishButtonText: fields.getTextValue(node, 'finish_button_text'),
     finishButtonUrl: fields.getTextValue(node, 'finish_button_url'),
     complete: async () => await completeLesson(course, lessonId),
   };